refactor(console): improve exports + move common functions into v2 entities
