+++ gravitee-apim-console-webui/src/entities/management-api-v2/index.ts
@@ -18,6 +18,7 @@ export * from './application';
 export * from './audit';
 export * from './plugin';
 export * from './createApi';
+export * from './documentation';
 export * from './event';
 export * from './group';
 export * from './log';
