+++ gravitee-apim-console-webui/src/management/api/documentation-v4/components/documentation-empty-state/api-documentation-v4-empty-state.component.ts
@@ -17,7 +17,7 @@
 import { Component, EventEmitter, OnDestroy, Output } from '@angular/core';
 import { Subject } from 'rxjs';
 
-import { getLogoForPageType, PageType } from '../../../../../entities/page';
+import { getLogoForPageType, PageType } from '../../../../../entities/management-api-v2';
 
 @Component({
   selector: 'api-documentation-empty-state',
@@ -37,5 +37,5 @@ export class ApiDocumentationV4EmptyStateComponent implements OnDestroy {
 
   // expose constants
   readonly getLogoForPageType = getLogoForPageType;
-  pageTypes = [PageType.MARKDOWN, PageType.SWAGGER, PageType.ASYNCAPI];
+  pageTypes: PageType[] = ['MARKDOWN', 'SWAGGER', 'ASYNCAPI'];
 }
