+++ gravitee-apim-console-webui/src/management/api/documentation-v4/components/documentation-list-navigation-header/api-documentation-v4-list-navigation-header.component.ts
@@ -16,8 +16,7 @@
 
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 
-import { Breadcrumb } from '../../../../../entities/management-api-v2/documentation/page';
-import { getLogoForPageType, PageType } from '../../../../../entities/page';
+import { Breadcrumb, getLogoForPageType, PageType } from '../../../../../entities/management-api-v2';
 
 @Component({
   selector: 'api-documentation-list-navigation-header',
@@ -38,5 +37,5 @@ export class ApiDocumentationV4ListNavigationHeaderComponent {
 
   // expose constants
   readonly getLogoForPageType = getLogoForPageType;
-  pageTypes = [PageType.MARKDOWN, PageType.SWAGGER, PageType.ASYNCAPI];
+  pageTypes: PageType[] = ['MARKDOWN', 'SWAGGER', 'ASYNCAPI'];
 }
